repos/promis-backend/requirements.txt
psycopg2
django==1.10.6
djangorestframework
django-extensions     # for graph models
markdown
django-filter
jsonfield
djsw_wrapper>=0.2.3.0
drf-nested-routers
dj-database-url
django-hvad
djangorestframework-gis
djangorestframework-filters
django-filters
pytz                  # for datetime conversions
django_loaddata_stdin # for loading test data from the host
pandas
